api/requirements.txt
django-cors-headers
requests
openai
